scanner fixed, appears in the right place but CSS done only at 20%
