+++ app/assets/stylesheets/components/_index.scss
@@ -3,12 +3,13 @@
 @import "avatar";
 @import "button";
 @import "cardsreview";
+@import "categories";
 @import "checkbox";
 @import "form";
 @import "main-container";
 @import "navbar";
+@import "scanner";
+@import "show";
 @import "written-reviewcard";
-@import "categories";
-@import "show"
