+++ app/controllers/products_controller.rb
@@ -12,8 +12,10 @@ class ProductsController < ApplicationController
   end
 
   def get_barcode
-
     @product = Product.find_or_initialize_by(upc: params[:upc])
+  end
+
+  def scan
   end
 
 end
