+++ app/javascript/packs/application.js
@@ -24,76 +24,13 @@ require("channels")
 
 // External imports
 import "bootstrap";
-import Quagga from 'quagga';
 
 // Internal imports, e.g:
-// import { initSelect2 } from '../components/init_select2';
+ import { initScanner } from '../plugins/init_scanner';
 
 document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
-  // initSelect2();
+  // initSelect2(); 
+  initScanner();
 });
 
-
-function order_by_occurrence(arr) {
-  var counts = {};
-  arr.forEach(function(value){
-      if(!counts[value]) {
-          counts[value] = 0;
-      }
-      counts[value]++;
-  });
-  return Object.keys(counts).sort(function(curKey,nextKey) {
-      return counts[curKey] < counts[nextKey];
-  });
-}
-
-function load_quagga(){
-  if ($('#barcode-scanner').length > 0 && navigator.mediaDevices && typeof navigator.mediaDevices.getUserMedia === 'function') {
-
-    var last_result = [];
-    if (Quagga.initialized == undefined) {
-      Quagga.onDetected(function(result) {
-        var last_code = result.codeResult.code;
-        last_result.push(last_code);
-
-        if (last_result.length > 20) {
-          const code = order_by_occurrence(last_result)[0];
-          last_result = [];
-          Quagga.stop();
-
-
-
-
-
-
-
-    var xmlHttp = new XMLHttpRequest();
-    xmlHttp.open( "GET", `/product/${code}`, true ); // false for synchronous request
-    xmlHttp.send( null );
-    console.log(code)
-    window.location.href = `/product/${code}`;
-
-        }
-      });
-    }
-
-    Quagga.init({
-      inputStream : {
-        name : "Live",
-        type : "LiveStream",
-        numOfWorkers: 1,
-        target: document.querySelector('#barcode-scanner')
-      },
-      decoder: {
-          readers : ['ean_reader']
-      }
-    },function(err) {
-        if (err) { console.log(err); return }
-        Quagga.initialized = true;
-        Quagga.start();
-    });
-
-  }
-};
-$(document).on('turbolinks:load', load_quagga);
