+++ app/views/products/scan.html.erb
@@ -1,0 +1,39 @@
+<div class='row'>
+    <div class='col-md-12'>
+      <div class="panel panel-primary">
+        <div class='panel-heading'>
+          <h3 class='panel-title'>Barcode scanner</h3>
+        </div>
+        <div class='panel-body'>
+          <div id='barcode-scanner'></div>
+        </div>
+      </div>
+    </div>
+</div>
+
+<div class='spacer'></div>
+  <div class='anim-box center spacer'>
+    <!-- empty first child used for corners -->
+    <div></div>
+    <div class='scanner'></div>
+    <div class='anim-item anim-item-sm'></div>
+    <div class='anim-item anim-item-lg'></div>
+    <div class='anim-item anim-item-lg'></div>
+    <div class='anim-item anim-item-sm'></div>
+    <div class='anim-item anim-item-lg'></div>
+    <div class='anim-item anim-item-sm'></div>
+    <div class='anim-item anim-item-md'></div>
+    <div class='anim-item anim-item-sm'></div>
+    <div class='anim-item anim-item-md'></div>
+    <div class='anim-item anim-item-lg'></div>
+    <div class='anim-item anim-item-sm'></div>
+    <div class='anim-item anim-item-sm'></div>
+    <div class='anim-item anim-item-lg'></div>
+    <div class='anim-item anim-item-sm'></div>
+    <div class='anim-item anim-item-lg'></div>
+    <div class='anim-item anim-item-sm'></div>
+    <div class='anim-item anim-item-lg'></div>
+    <div class='anim-item anim-item-sm'></div>
+    <div class='anim-item anim-item-md'></div> 
+  </div>
+  <div class='spacer'></div>
