+++ app/views/shared/_navbar.html.erb
@@ -2,14 +2,14 @@
   <!-- inactive links scan/favorites/user -->
   <%= link_to root_path, class: "navbar-brand a-navbar" do %>
     <%= image_tag("icons/home_icon.png") %>
-    <% end %>
-  <%= link_to "scan", class: "navbar-brand a-navbar" do %>
+  <% end %>
+  <%= link_to scan_path, class: "navbar-brand a-navbar" do %>
     <%= image_tag("icons/scan_icon.png") %>
-    <% end %>
+  <% end %>
   <%= link_to "favorites", class: "navbar-brand a-navbar" do %>
     <%= image_tag("icons/heart_icon.png") %>
-    <% end %>
+  <% end %>
   <%= link_to "user", class: "navbar-brand a-navbar" do %>
     <%= image_tag("icons/profile_icon.png") %>
-    <% end %>
+  <% end %>
 </div>
